refactor(#97): graduate Save{Error,Success}Snackbar to src/components/feedback/
Both files are 100% generic (no vehicle-specific imports, generic props)
and already had two call sites — moving them out of the feature dir
removes the cross-feature leak that any future sidebar-edit consumer
would otherwise create by importing from data/vehicles/components/.

git mv:
- data/vehicles/components/SaveErrorSnackbar.tsx → components/feedback/
- data/vehicles/components/SaveSuccessSnackbar.tsx → components/feedback/

Imports updated: VehicleDetails, VehicleCreatePage.
